+++ src/test/java/com/jaw/menu/application/MenuGroupServiceTest.java
@@ -90,7 +90,7 @@ class MenuGroupServiceTest {
         return Menu.builder()
             .name(name)
             .englishName(englishName)
-            .price(BigDecimal.valueOf(price))
+            .price(price)
             .menuGroup(menuGroup)
             .build();
     }
